[Refactor] Tag 도메인 분리
